+++ libraries/arrays/src/__tests__/toPickWeightedItemFromDistribution.unit.test.ts
@@ -5,7 +5,7 @@ describe('toPickFromDistribution()', () => {
   it('returns a function that picks an item based on weights', () => {
     const items = ['a', 'b', 'c'];
     const distribution = { 'a': 1, 'b': 2, 'c': 3 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);
 
@@ -17,7 +17,7 @@ describe('toPickFromDistribution()', () => {
   it('picks an item using a custom default weight', () => {
     const items = ['a', 'b', 'c'];
     const distribution = { 'a': 0, 'b': 0 };
-    const options = { defaultWeight: 1, getName: (item: string) => item };
+    const options = { defaultWeight: 1, getCategory: (item: string) => item };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);
     const pickedItem = pickItem();
@@ -29,7 +29,7 @@ describe('toPickFromDistribution()', () => {
   it('excludes items with zero or negative weights', () => {
     const items = ['a', 'b', 'c'];
     const distribution = { 'a': 1, 'b': 0, 'c': -1 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);
     const pickedItem = pickItem();
@@ -40,7 +40,7 @@ describe('toPickFromDistribution()', () => {
   it('uses the accessor function to get the name of the item', () => {
     const items = [{ name: 'a' }, { name: 'b' }, { name: 'c' }];
     const distribution = { 'a': 1, 'b': 2, 'c': 3 };
-    const options = { getName: (item: { name: string }) => item.name };
+    const options = { getCategory: (item: { name: string }) => item.name };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);
     const pickedItem = pickItem();
@@ -51,7 +51,7 @@ describe('toPickFromDistribution()', () => {
   it('accepts read-only arrays', () => {
     const items = Object.freeze(['a', 'b', 'c']);
     const distribution = { 'a': 1, 'b': 2, 'c': 3 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);
 
@@ -62,7 +62,7 @@ describe('toPickFromDistribution()', () => {
   it('throws an error if all weights are zero or negative', () => {
     const items = ['a', 'b', 'c'];
     const distribution = { 'a': 0, 'b': -1, 'c': -2 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
 
     const throwingFn = () => toPickWeightedItemFromDistribution(items, distribution, options);
 
@@ -76,7 +76,7 @@ describe('toPickFromDistribution()', () => {
   it('throws an error if the items array is empty', () => {
     const items: string[] = [];
     const distribution = { 'a': 1, 'b': 2, 'c': 3 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
 
     const throwingFn = () => toPickWeightedItemFromDistribution(items, distribution, options);
 
@@ -90,7 +90,7 @@ describe('toPickFromDistribution()', () => {
   it('accepts and uses options when picking an item', () => {
     const items = ['a', 'b', 'c'];
     const distribution = { 'a': 1, 'b': 2, 'c': 3 };
-    const options = { getName: (item: string) => item };
+    const options = { getCategory: (item: string) => item };
     const pickOptions = { seed: 1234 };
 
     const pickItem = toPickWeightedItemFromDistribution(items, distribution, options);