+++ libraries/arrays/src/toPickWeightedItemFromDistribution.ts
@@ -5,18 +5,18 @@ import { toPickWeightedItem } from './toPickWeightedItem.ts';
  * Returns a function that pseudo-randomly picks an item from the array using weighted odds, which are derived
  * from the provided distribution.
  */
-export function toPickWeightedItemFromDistribution<T>(
-  items: ReadonlyArray<T>,
-  distribution: Distribution,
-  params: Params<T>,
-): ReturnType<typeof toPickWeightedItem<T>> {
-  const { defaultWeight = 0, getName } = params;
+export function toPickWeightedItemFromDistribution<TItem, TCategory extends string>(
+  items: ReadonlyArray<TItem>,
+  distribution: Partial<Distribution<TCategory>>,
+  params: Params<TItem, TCategory>,
+): ReturnType<typeof toPickWeightedItem<TItem>> {
+  const { defaultWeight = 0, getCategory } = params;
 
-  const filteredItems: T[] = [];
+  const filteredItems: TItem[] = [];
   const filteredWeights: number[] = [];
 
   items.forEach((item) => {
-    const weight = distribution[getName(item)] ?? defaultWeight;
+    const weight = distribution[getCategory(item)] ?? defaultWeight;
     if (weight > 0) {
       filteredItems.push(item);
       filteredWeights.push(weight);
@@ -32,11 +32,9 @@ export function toPickWeightedItemFromDistribution<T>(
   return toPickWeightedItem(filteredItems, cumulativeWeights);
 }
 
-interface Distribution {
-  [key: string]: number;
-}
+type Distribution<TCategory extends string> = Record<TCategory, number>;
 
-interface Params<T> {
+interface Params<TItem, TCategory> {
   defaultWeight?: number;
-  getName: (item: T) => string;
+  getCategory: (item: TItem) => TCategory;
 }